Renames reportback helper as util
