+++ lib/helpers/campaignActivity.js
@@ -4,7 +4,7 @@ const superagent = require('superagent');
 const logger = require('winston');
 const requestHelper = require('./request');
 const rogue = require('../rogue');
-const util = require('./reportback');
+const util = require('./util');
 const rogueClientConfig = require('../../config/lib/rogue/rogue-client');
 
 /**
