+++ lib/helpers/index.js
@@ -5,8 +5,8 @@ const cache = require('./cache');
 const campaign = require('./campaign');
 const campaignActivity = require('./campaignActivity');
 const command = require('./command');
-const reportback = require('./reportback');
 const request = require('./request');
+const util = require('./util');
 
 module.exports = {
   botConfig,
@@ -14,6 +14,6 @@ module.exports = {
   campaign,
   campaignActivity,
   command,
-  reportback,
   request,
+  util,
 };
