+++ lib/helpers/request.js
@@ -2,7 +2,7 @@
 
 const logger = require('winston');
 const command = require('./command');
-const util = require('./reportback');
+const util = require('./util');
 
 /**
  * @return {String}
