+++ lib/helpers/util.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const config = require('../../config/lib/helpers/reportback');
+const config = require('../../config/lib/helpers/util');
 
 /**
  * @param {string} message
